requirements.txt
python-dotenv
SQLAlchemy>=2.0
alembic
celery[redis]
jinja2
google-api-python-client
pytest
structlog